requirements.txt
Flask==3.0.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2
openpyxl==3.1.2
requests==2.31.0
pyserial==3.5
python-dotenv==1.0.0
